requirements.txt
# Python dependencies for cloud engineering scripts

# AWS SDK
boto3>=1.34.0

# Google Cloud SDKs
google-cloud-compute>=1.16.0
google-cloud-storage>=2.14.0

# Azure SDKs
azure-identity>=1.15.0
azure-mgmt-compute>=30.5.0
azure-mgmt-storage>=21.1.0

# Additional utilities (if needed in future)
# requests>=2.31.0
# python-dotenv>=1.0.0  # Alternative to manual .env parsing
